Updated build scripts to newest library versions

+++ getlibs.sh
@@ -24,6 +24,17 @@ elif [ "$CMAKE_BUILD_TYPE" != Release ]; then
     echo "WARNING: building libs with CMAKE_BUILD_TYPE=$CMAKE_BUILD_TYPE"
 fi
 
+if [ -n "${CMAKE_TOOLCHAIN_FILE+x}" ]; then
+    TOOLCHAIN_FILE="$CMAKE_TOOLCHAIN_FILE"
+fi
+
+function show_cmake_vars() {
+    for var in CMAKE_TOOLCHAIN_FILE CMAKE_BUILD_TYPE CMAKE_PREFIX_PATH CMAKE_INSTALL_PREFIX \
+               CMAKE_GENERATOR CMAKE_BUILD_PARALLEL_LEVEL; do
+        echo "  $var=${!var:- (unset)}"
+    done
+}
+
 for target in $targets ; do
     echo "GETLIBS for $target"
     BUILD_TARGET=$target
@@ -32,6 +43,9 @@ for target in $targets ; do
     PREFIX="$PROJECT_DIR/local/${BUILD_TARGET}"
     # export CMAKE_PREFIX_PATH="$PREFIX"
     export CMAKE_INSTALL_PREFIX="$PREFIX"
+    if [ -z "${TOOLCHAIN_FILE+x}" ] && [ -r "$PROJECT_DIR/cmake/toolchain/${target}.cmake" ] ; then
+        export CMAKE_TOOLCHAIN_FILE="$PROJECT_DIR/cmake/toolchain/${target}.cmake"
+    fi
 
     ## DOCTEST
     NAME=doctest
@@ -50,10 +64,7 @@ for target in $targets ; do
         [ -d "${SOURCE}" ] || tar xf "${ARCHIVE}"
         popd
         echo "Building $LIBRARY in $BUILD from $SOURCE"
-        echo "  CMAKE_TOOLCHAIN_FILE=${CMAKE_TOOLCHAIN_FILE:-(unset)}"
-        echo "  CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE:-(unset)}"
-        echo "  CMAKE_PREFIX_PATH=${CMAKE_PREFIX_PATH:-(unset)}"
-        echo "  CMAKE_INSTALL_PREFIX=${CMAKE_INSTALL_PREFIX:-(unset)}"
+        show_cmake_vars
         cmake -S "$SOURCE" -B "$BUILD" -DDOCTEST_WITH_TESTS=OFF -DDOCTEST_WITH_MAIN_IN_STATIC_LIB=ON
         cmake --build "$BUILD" --config $CMAKE_BUILD_TYPE
         cmake --install "$BUILD" --config $CMAKE_BUILD_TYPE --prefix "${CMAKE_INSTALL_PREFIX}"
@@ -63,10 +74,10 @@ for target in $targets ; do
 
     ## XXHASH (in src/kernel.c)
     NAME=xxHash
-    VERSION=0.8.2
+    VERSION=0.8.3
     LIBRARY="${NAME}-${VERSION}"
     ARCHIVE="$LIBRARY.tgz"
-    SHA256=baee0c6afd4f03165de7a4e67988d16f0f2b257b51d0e3cb91909302a26a79c4
+    SHA256=aae608dfe8213dfd05d909a57718ef82f30722c392344583d3f39050c7f29a80
     SOURCE="${SOURCES}/$LIBRARY"
     BUILD="${PREFIX}/build-$LIBRARY"
     if [ -r "${CMAKE_INSTALL_PREFIX}/include/xxhash.h" ]; then
@@ -78,11 +89,37 @@ for target in $targets ; do
         [ -d "$SOURCE" ] || tar xf "${ARCHIVE}"
         popd
         echo "Building $LIBRARY in $BUILD from $SOURCE"
-        echo "  CMAKE_TOOLCHAIN_FILE=${CMAKE_TOOLCHAIN_FILE:-(unset)}"
-        echo "  CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE:-(unset)}"
-        echo "  CMAKE_PREFIX_PATH=${CMAKE_PREFIX_PATH:-(unset)}"
-        echo "  CMAKE_INSTALL_PREFIX=${CMAKE_INSTALL_PREFIX:-(unset)}"
+        show_cmake_vars
         cmake -S "$SOURCE/cmake_unofficial" -B "$BUILD" -DBUILD_SHARED_LIBS=OFF
+        cmake --build "$BUILD" --config $CMAKE_BUILD_TYPE
+        cmake --install "$BUILD" --config $CMAKE_BUILD_TYPE --prefix "${CMAKE_INSTALL_PREFIX}"
+        rm -Rf "$BUILD"
+        rm -Rf "$SOURCE"
+    fi
+
+    ## BOOST for UUtils
+    NAME=boost
+    VERSION=1.88.0
+    LIBRARY="${NAME}-${VERSION}"
+    ARCHIVE="${LIBRARY}-cmake.tar.xz"
+    SHA256=f48b48390380cfb94a629872346e3a81370dc498896f16019ade727ab72eb1ec
+    SOURCE="${SOURCES}/${LIBRARY}"
+    BUILD="${PREFIX}/build-${LIBRARY}"
+    if [ -r "${CMAKE_INSTALL_PREFIX}/include/boost/math/distributions/arcsine.hpp" ] ; then
+        echo "$LIBRARY is already installed in $CMAKE_INSTALL_PREFIX"
+    else
+        pushd "$SOURCES"
+        [ -r "${ARCHIVE}" ] || curl -sL "https://github.com/boostorg/boost/releases/download/${LIBRARY}/${ARCHIVE}" -o "${ARCHIVE}"
+        if [ -n "$(command -v shasum)" ]; then
+            echo "$SHA256  $ARCHIVE" | shasum -a256 --check -
+        fi
+        [ -d "${SOURCE}" ] || tar xf "${ARCHIVE}"
+        popd
+        echo "Building $LIBRARY in $BUILD from $SOURCE"
+        show_cmake_vars
+        cmake -S "$SOURCE" -B "$BUILD" -DBUILD_SHARED_LIBS=OFF \
+          -DBOOST_INCLUDE_LIBRARIES="headers;math" -DBOOST_ENABLE_MPI=OFF -DBOOST_ENABLE_PYTHON=OFF \
+          -DBOOST_RUNTIME_LINK=static -DBUILD_TESTING=OFF -DBOOST_INSTALL_LAYOUT=system
         cmake --build "$BUILD" --config $CMAKE_BUILD_TYPE
         cmake --install "$BUILD" --config $CMAKE_BUILD_TYPE --prefix "${CMAKE_INSTALL_PREFIX}"
         rm -Rf "$BUILD"
@@ -91,10 +128,10 @@ for target in $targets ; do
 
     # UUtils various low level Uppaal utilities
     NAME=UUtils
-    VERSION=2.0.5
+    VERSION=2.0.7
     LIBRARY="${NAME}-${VERSION}"
     ARCHIVE="${LIBRARY}.tar.gz"
-    SHA256=e213f936af73344de071a7794233a328028045c08df58ac9c637a0e6a2ad7b3f
+    SHA256=52a823768398707eee42392182c320141662e74e6bd8eaac1a0eca22d8a27bcd
     SOURCE="${SOURCES}/${LIBRARY}"
     BUILD="${PREFIX}/build-${LIBRARY}"
     if [ -r "$PREFIX/include/base/Enumerator.h" ]; then
@@ -107,11 +144,7 @@ for target in $targets ; do
       [ -d "$SOURCE" ] || tar -xf "$ARCHIVE"
       popd
       echo "Building $LIBRARY in $BUILD"
-      echo "  CMAKE_TOOLCHAIN_FILE=${CMAKE_TOOLCHAIN_FILE:-(unset)}"
-      echo "  CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE:-(unset)}"
-      echo "  CMAKE_PREFIX_PATH=${CMAKE_PREFIX_PATH:-(unset)}"
-      echo "  CMAKE_INSTALL_PREFIX=${CMAKE_INSTALL_PREFIX:-(unset)}"
-      echo "  CMAKE_GENERATOR=${CMAKE_GENERATOR:-(unset)}"
+      show_cmake_vars
       cmake -S "$SOURCE" -B "$BUILD" -DUUtils_WITH_TESTS=OFF -DUUtils_WITH_BENCHMARKS=OFF
       cmake --build "$BUILD" --config $CMAKE_BUILD_TYPE
       cmake --install "$BUILD" --config $CMAKE_BUILD_TYPE --prefix="$CMAKE_INSTALL_PREFIX"
@@ -121,10 +154,10 @@ for target in $targets ; do
 
     # UDBM Uppaal Difference Bound Matrix library
     NAME=UDBM
-    VERSION=2.0.14
+    VERSION=2.0.15
     LIBRARY="${NAME}-${VERSION}"
     ARCHIVE="${LIBRARY}.tar.gz"
-    SHA256=019ac8ea4ca52ebd4d4aedf50f3a039648776b62b854534e0c678004fec25b08
+    SHA256=56a630731f05ef8c2c7122ca2ca225f7608e1ff1efb4b980f0ee6bdcc388efa5
     SOURCE="${SOURCES}/${LIBRARY}"
     BUILD="${PREFIX}/build-${LIBRARY}"
     if [ -r "$PREFIX/include/dbm/config.h" ]; then
@@ -137,12 +170,8 @@ for target in $targets ; do
       [ -d "$SOURCE" ] || tar -xf "$ARCHIVE"
       popd
       echo "Building $LIBRARY in $BUILD"
-      echo "  CMAKE_TOOLCHAIN_FILE=${CMAKE_TOOLCHAIN_FILE:-(unset)}"
-      echo "  CMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE:-(unset)}"
-      echo "  CMAKE_PREFIX_PATH=${CMAKE_PREFIX_PATH:-(unset)}"
-      echo "  CMAKE_INSTALL_PREFIX=${CMAKE_INSTALL_PREFIX:-(unset)}"
-      echo "  CMAKE_GENERATOR=${CMAKE_GENERATOR:-(unset)}"
-      cmake -S "$SOURCE" -B "$BUILD" -DUDBM_WITH_TESTS=OFF
+      show_cmake_vars
+      cmake -S "$SOURCE" -B "$BUILD" -DUDBM_WITH_TESTS=OFF -DUDBM_CLANG_TIDY=OFF
       cmake --build "$BUILD" --config $CMAKE_BUILD_TYPE
       cmake --install "$BUILD" --config $CMAKE_BUILD_TYPE --prefix="$CMAKE_INSTALL_PREFIX"
       rm -Rf "$BUILD"
